+++ src/main/java/com/atlassian/performance/tools/sshubuntu/api/SshUbuntu.java
@@ -1,5 +1,9 @@
 package com.atlassian.performance.tools.sshubuntu.api;
 
+import org.testcontainers.containers.GenericContainer;
+
 public interface SshUbuntu extends AutoCloseable {
     SshHost getSsh();
+
+    GenericContainer getContainer();
 }